+++ IBAnimatable/PortalAnimator.swift
@@ -27,12 +27,14 @@ public class PortalAnimator: NSObject, AnimatedTransitioning {
     }
     
     switch fromDirection {
-    case .Backward:
-      self.transitionAnimationType = .Portal(direction: .Backward, params: params)
-      self.reverseAnimationType = .Portal(direction: .Forward, params: params)
-    default:
+    case .Forward:
       self.transitionAnimationType = .Portal(direction: .Forward, params: params)
       self.reverseAnimationType = .Portal(direction: .Backward, params: params)
+      self.interactiveGestureType = .Pinch(direction: .Close)
+    default:
+      self.transitionAnimationType = .Portal(direction: .Backward, params: params)
+      self.reverseAnimationType = .Portal(direction: .Forward, params: params)
+      self.interactiveGestureType = .Pinch(direction: .Open)
     }
 
     super.init()
@@ -51,10 +53,11 @@ extension PortalAnimator: UIViewControllerAnimatedTransitioning {
       return
     }
     
-    if fromDirection == .Forward {
+    switch fromDirection {
+    case .Forward:
       executeForwardAnimation(transitionContext, containerView: containerView, fromView: fromView, toView: toView)
-    } else {
-      executeBackwardAnimations(transitionContext, containerView: containerView, fromView: fromView, toView: toView)
+    default:
+      executeBackwardAnimation(transitionContext, containerView: containerView, fromView: fromView, toView: toView)
     }
   }
   
@@ -80,24 +83,27 @@ private extension PortalAnimator {
     rightHandView.frame = rightSnapshotRegion
     containerView.addSubview(rightHandView)
     
-    fromView.removeFromSuperview()
+    fromView.hidden = true
 
-    UIView.animateWithDuration(transitionDuration, delay: 0.0, options: .CurveEaseOut, animations: {
-      leftHandView.frame = leftHandView.frame.offsetBy(dx: -leftHandView.frame.width, dy: 0.0)
-      rightHandView.frame = rightHandView.frame.offsetBy(dx: rightHandView.frame.width, dy: 0.0)
-      toViewSnapshot.center = toView.center
-      toViewSnapshot.frame = containerView.frame
-    }, completion: { _ in
-      if transitionContext.transitionWasCancelled() {
-        containerView.addSubview(fromView)
-        self.removeOtherViews(fromView)
-      } else {
-        toView.frame = containerView.frame
-        containerView.addSubview(toView)
-        self.removeOtherViews(toView)
+    UIView.animateWithDuration(transitionDuration, delay: 0.0, options: .CurveEaseOut,
+      animations: {
+        leftHandView.frame = leftHandView.frame.offsetBy(dx: -leftHandView.frame.width, dy: 0.0)
+        rightHandView.frame = rightHandView.frame.offsetBy(dx: rightHandView.frame.width, dy: 0.0)
+        toViewSnapshot.center = toView.center
+        toViewSnapshot.frame = containerView.frame
+      },
+      completion: { _ in
+        fromView.hidden = false
+        if transitionContext.transitionWasCancelled() {
+          self.removeOtherViews(fromView)
+        } else {
+          toView.frame = containerView.frame
+          containerView.addSubview(toView)
+          self.removeOtherViews(toView)
+        }
+        transitionContext.completeTransition(!transitionContext.transitionWasCancelled())
       }
-      transitionContext.completeTransition(!transitionContext.transitionWasCancelled())
-    })
+    )
   }
   
 }
@@ -106,7 +112,7 @@ private extension PortalAnimator {
 
 private extension PortalAnimator {
   
-  func executeBackwardAnimations(transitionContext: UIViewControllerContextTransitioning, containerView: UIView, fromView: UIView, toView: UIView) {
+  func executeBackwardAnimation(transitionContext: UIViewControllerContextTransitioning, containerView: UIView, fromView: UIView, toView: UIView) {
     containerView.addSubview(fromView)
     toView.frame = toView.frame.offsetBy(dx: toView.frame.width, dy: 0)
     containerView.addSubview(toView)
@@ -123,22 +129,24 @@ private extension PortalAnimator {
     rightHandView.frame = rightHandView.frame.offsetBy(dx: rightHandView.frame.width, dy: 0)
     containerView.addSubview(rightHandView)
 
-    UIView.animateWithDuration(transitionDuration, delay: 0.0, options: .CurveEaseOut, animations: {
-      leftHandView.frame = leftHandView.frame.offsetBy(dx: leftHandView.frame.size.width, dy: 0)
-      rightHandView.frame = rightHandView.frame.offsetBy(dx: -rightHandView.frame.size.width, dy: 0)
-      let scale = CATransform3DIdentity
-      fromView.layer.transform = CATransform3DScale(scale, self.zoomScale, self.zoomScale, 1)
-    }, completion: { _ in
-      
-      if transitionContext.transitionWasCancelled() {
-        self.removeOtherViews(fromView)
-      } else {
-        self.removeOtherViews(toView)
-        toView.frame = containerView.bounds
-        fromView.layer.transform = CATransform3DIdentity
+    UIView.animateWithDuration(transitionDuration, delay: 0.0, options: .CurveEaseOut,
+      animations: {
+        leftHandView.frame = leftHandView.frame.offsetBy(dx: leftHandView.frame.size.width, dy: 0)
+        rightHandView.frame = rightHandView.frame.offsetBy(dx: -rightHandView.frame.size.width, dy: 0)
+        let scale = CATransform3DIdentity
+        fromView.layer.transform = CATransform3DScale(scale, self.zoomScale, self.zoomScale, 1)
+      },
+      completion: { _ in
+        if transitionContext.transitionWasCancelled() {
+          self.removeOtherViews(fromView)
+        } else {
+          self.removeOtherViews(toView)
+          toView.frame = containerView.bounds
+          fromView.layer.transform = CATransform3DIdentity
+        }
+        transitionContext.completeTransition(!transitionContext.transitionWasCancelled())
       }
-      transitionContext.completeTransition(!transitionContext.transitionWasCancelled())
-    })
+    )
   }
   
 }